GSDX: Move PSX code to an isolated path
