+++ plugins/GSdx/CMakeLists.txt
@@ -39,14 +39,14 @@ endif()
 set(GSdxSources
     GLLoader.cpp
     GLState.cpp
-    GPU.cpp
-    GPUDrawScanline.cpp
-    GPUDrawScanlineCodeGenerator.cpp
-    GPULocalMemory.cpp
-    GPURenderer.cpp
-    GPURendererSW.cpp
-    GPUSetupPrimCodeGenerator.cpp
-    GPUState.cpp
+    PSX/GPU.cpp
+    PSX/GPUDrawScanline.cpp
+    PSX/GPUDrawScanlineCodeGenerator.cpp
+    PSX/GPULocalMemory.cpp
+    PSX/GPURenderer.cpp
+    PSX/GPURendererSW.cpp
+    PSX/GPUSetupPrimCodeGenerator.cpp
+    PSX/GPUState.cpp
     GS.cpp
     GSAlignedClass.cpp
     GSBlock.cpp
@@ -72,7 +72,7 @@ set(GSdxSources
     GSHwHack.cpp
     GSLinuxDialog.cpp
     GSLocalMemory.cpp
-	GSLzma.cpp
+    GSLzma.cpp
     GSPerfMon.cpp
     GSPng.cpp
     GSRasterizer.cpp
@@ -111,17 +111,17 @@ set(GSdxSources
     )
 
 set(GSdxHeaders
-    GPU.h
-    GPUDrawScanline.h
-    GPUDrawScanlineCodeGenerator.h
-    GPUDrawingEnvironment.h
-    GPULocalMemory.h
-    GPURenderer.h
-    GPURendererSW.h
-    GPUScanlineEnvironment.h
-    GPUSetupPrimCodeGenerator.h
-    GPUState.h
-    GPUVertex.h
+    PSX/GPU.h
+    PSX/GPUDrawScanline.h
+    PSX/GPUDrawScanlineCodeGenerator.h
+    PSX/GPUDrawingEnvironment.h
+    PSX/GPULocalMemory.h
+    PSX/GPURenderer.h
+    PSX/GPURendererSW.h
+    PSX/GPUScanlineEnvironment.h
+    PSX/GPUSetupPrimCodeGenerator.h
+    PSX/GPUState.h
+    PSX/GPUVertex.h
     GS.h
     GSAlignedClass.h
     GSBlock.h
