+++ plugins/GSdx/PSX/GPU.cpp
@@ -22,8 +22,8 @@
 #include "stdafx.h"
 #include "GSdx.h"
 #include "GSUtil.h"
-#include "GPURendererSW.h"
 #include "GSDeviceNull.h"
+#include "GPURendererSW.h"
 
 #ifdef _WIN32
