+++ plugins/GSdx/PSX/GPUDrawingEnvironment.h
