+++ plugins/GSdx/PSX/GPURenderer.h
