+++ plugins/GSdx/PSX/GPURendererSW.cpp
