+++ plugins/GSdx/PSX/GPUScanlineEnvironment.h
