+++ plugins/GSdx/PSX/GPUSettingsDlg.h
