+++ plugins/GSdx/PSX/GPUSetupPrimCodeGenerator.cpp
