+++ plugins/GSdx/PSX/GPUSetupPrimCodeGenerator.h
