+++ plugins/GSdx/PSX/GPUVertex.h
